refactor: taste review fixes and provider API key display
- Split proxy-core.ts (689→335 lines) into upstream-call.ts, log-helpers.ts,
  enhancement-handler.ts for single-responsibility
- Fix 3 silent catches with proper logging (proxy enhancement config,
  allowed_models, model-info injection)
- Extract inline enhancement block to applyEnhancement() function
- Add PUT validation for proxy-enhancement endpoint
- Fix model-state.ts magic numbers (extract HOUR_MS constant)
- Add loading state to ProxyEnhancement.vue, use Checkbox instead of Switch
- Fix proxy-enhancement route path (missing /admin/api prefix)
- Show full decrypted API key in provider list and edit form
- Extract RequestLogInsert interface in logs.ts
- Update tests to match new api_key response behavior
