+++ frontend/src/views/Providers.vue
@@ -102,8 +102,8 @@
             <Input v-model="form.base_url" type="url" required />
           </div>
           <div>
-            <Label class="block text-sm font-medium text-foreground mb-1">API Key {{ editingId ? '(留空不修改)' : '' }}</Label>
-            <Input v-model="form.api_key" :type="editingId ? 'password' : 'text'" :required="!editingId" />
+            <Label class="block text-sm font-medium text-foreground mb-1">API Key</Label>
+            <Input v-model="form.api_key" type="text" required />
           </div>
           <div>
             <Label class="block text-sm font-medium text-foreground mb-1">可用模型</Label>
@@ -239,7 +239,7 @@ function openCreate() {
 
 function openEdit(p: Provider) {
   editingId.value = p.id
-  form.value = { name: p.name, api_type: p.api_type, base_url: p.base_url, api_key: '', models: [...(p.models || [])], is_active: !!p.is_active }
+  form.value = { name: p.name, api_type: p.api_type, base_url: p.base_url, api_key: p.api_key, models: [...(p.models || [])], is_active: !!p.is_active }
   modelInput.value = ''
   presetGroup.value = ''
   presetPlan.value = ''
